More VC++/Win32 fixes.

+++ dangerdeep/src/model.h
@@ -279,7 +279,7 @@ public:
 		/// construct a voxel
 		voxel(const vector3f& rp, float pv, float m, float rv)
 			: relative_position(rp), part_of_volume(pv),
-			root3_part_of_volume(pow(pv, 1.0/3.0)), relative_mass(m),
+			root3_part_of_volume(pow(pv, (float)(1.0/3.0))), relative_mass(m),
 			relative_volume(rv)
 		{
 			for (int i = 0; i < 6; ++i) neighbour_idx[i] = -1;
